test(taiko-client-rs): consolidate whitelist test fixtures and mocks
