+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/codec.rs
@@ -347,65 +347,9 @@ pub(crate) fn decompress_tx_list(bytes: &[u8]) -> Result<Vec<u8>> {
 }
 
 #[cfg(test)]
-pub(crate) mod tests {
-    use alloy_primitives::{Bloom, Bytes};
-
+mod tests {
     use super::*;
-
-    /// Reads a Go-generated fixture, failing with a regeneration hint.
-    pub(crate) fn go_fixture(rel: &str) -> Vec<u8> {
-        let path = format!("{}/fixtures/go/{rel}", env!("CARGO_MANIFEST_DIR"));
-        std::fs::read(&path)
-            .unwrap_or_else(|e| panic!("missing fixture {path} ({e}) — run `just gen-fixtures`"))
-    }
-
-    /// Parses a fixture JSON sidecar.
-    pub(crate) fn go_fixture_json(rel: &str) -> serde_json::Value {
-        serde_json::from_slice(&go_fixture(rel)).expect("fixture json")
-    }
-
-    /// Signs a 32-byte prehash with the golden-touch [`FixedKSigner`] and returns
-    /// the 65-byte wire signature (`r || s || recovery_id`).
-    ///
-    /// Shared crate-wide test helper: mirrors the production
-    /// `payload_build::sign_digest` assembly so callers get a signature that
-    /// [`recover_signer`] accepts without assuming any v-byte convention.
-    pub(crate) fn fixed_k_sign(prehash: B256) -> [u8; SIGNATURE_LEN] {
-        let signer = protocol::FixedKSigner::golden_touch().expect("golden touch signer");
-        let sig = signer.sign_with_predefined_k(prehash.as_ref()).expect("fixed-k signature");
-
-        let mut sig_bytes = [0u8; SIGNATURE_LEN];
-        sig_bytes[..32].copy_from_slice(&sig.signature.r().to_be_bytes::<32>());
-        sig_bytes[32..64].copy_from_slice(&sig.signature.s().to_be_bytes::<32>());
-        sig_bytes[64] = sig.recovery_id;
-        sig_bytes
-    }
-
-    pub(crate) fn sample_envelope() -> WhitelistExecutionPayloadEnvelope {
-        WhitelistExecutionPayloadEnvelope {
-            end_of_sequencing: Some(true),
-            is_forced_inclusion: Some(true),
-            parent_beacon_block_root: Some(B256::from([0xabu8; 32])),
-            header_difficulty: Some(U256::from(0x12345678u64)),
-            execution_payload: ExecutionPayloadV1 {
-                parent_hash: B256::from([0x01u8; 32]),
-                fee_recipient: Address::from([0x11u8; 20]),
-                state_root: B256::from([0x02u8; 32]),
-                receipts_root: B256::from([0x03u8; 32]),
-                logs_bloom: Bloom::default(),
-                prev_randao: B256::from([0x04u8; 32]),
-                block_number: 42,
-                gas_limit: 30_000_000,
-                gas_used: 21000,
-                timestamp: 1_735_000_000,
-                extra_data: Bytes::from(vec![0x55u8; 8]),
-                base_fee_per_gas: U256::from(1_000_000_000u64),
-                block_hash: B256::from([0x05u8; 32]),
-                transactions: vec![Bytes::from(vec![0x99u8; 4])],
-            },
-            signature: Some([0x22u8; SIGNATURE_LEN]),
-        }
-    }
+    use crate::test_support::{go_fixture, go_fixture_json, sample_envelope};
 
     #[test]
     fn encode_unsafe_request_is_raw_hash_bytes() {
